move timer0 to its own module
