+++ RobotNoASF/Interfaces/imu_interface.h
@@ -73,27 +73,7 @@ typedef struct euler_packet
 */
 int imuInit(void);
 
-/*
-* Function:
-* void timer0Init(void)
-*
-* initialise timer0. will be moved to its own module soon
-*
-* Inputs:
-* none
-*
-* Returns:
-* none
-*
-* Implementation:
-* TODO:[explain key steps of function] timer0 init
-* [use heavy detail for anything complicated]
-*
-* Improvements:
-* Move to its own module.
-*
-*/
-void timer0Init(void);
+
 
 /*
 * Function: int imuDmpInit(void)
@@ -189,35 +169,6 @@ unsigned short invRow2Scale(const signed char *row);
 */
 void getEulerAngles(long *ptQuat, euler_packet_t *eulerAngle);
 
-/*
-* Function: int get_ms(uint32_t *timestamp)
-*
-* Required by the IMU drivers (hence naming convention). Outputs the system uptime generated from
-* Timer0.
-*
-* Inputs:
-* address of an integer where the timestamp will be stored
-*
-* Returns:
-* function will return 1 if invalid pointer is passed, otherwise a 0 on success
-*
-*/
-int get_ms(uint32_t *timestamp);
-
-/*
-* Function: int delay_ms(uint32_t period_ms)
-*
-* Required by the IMU drivers (hence naming convention). Halts execution for desired number of
-* milliseconds.
-*
-* Inputs:
-* period_ms is the number of milliseconds to wait
-*
-* Returns:
-* Always returns 0
-*
-*/
-int delay_ms(uint32_t period_ms);
 
 /*
 * Function: char twiWriteImu(unsigned char slave_addr, unsigned char reg_addr,
@@ -275,21 +226,6 @@ char twiReadImu(unsigned char slave_addr, unsigned char reg_addr,
 */
 uint8_t imuCommTest(void);
 
-/*
-* Function: void TC0_Handler()
-*
-* Interrupt handler for Timer0. Is used to help implement get_ms() and delay_ms() functions
-* required by the IMU driver. Is also used to trigger reading the IMU's FIFO buffer (until
-* hardware interrupts are implemented). The only interrupt on Timer0 is on Register C compare,
-* which will trigger an interrupt once every millisecond
-*
-* Inputs:
-* none
-*
-* Returns:
-* Increments systemTimestamp once every millisecond.
-*
-*/
-void TC0_Handler();
+
 
 #endif /* IMU_INTERFACE_H_ */
